+++ src/hooks/index.ts
@@ -1,5 +1,4 @@
 export * from './useAutoSave';
-export * from './useDatasources';
 export * from './useFormLayout';
 export * from './useMutableState';
 export * from './useQueryFields';