+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added form-element of type color (#561)
 - Updated new lines in Text Area and Code Editor (#563)
 - Added support variables in data sources (#564)
+- Updated packages for Code Editor (#569)
 - Updated element helpers (#570)
 
 ## 4.9.0 (2024-11-16)